Refactor the report run algorithm (issue #14)
Closes #14
