+++ utils/exec.go
@@ -4,19 +4,31 @@ import (
 	"bufio"
 	"bytes"
 	"fmt"
-	log "github.com/Sirupsen/logrus"
 	"io"
+	"io/ioutil"
 	"os"
 	"os/exec"
 	"runtime"
 	"strings"
 	"syscall"
 	"time"
+
+	log "github.com/Sirupsen/logrus"
 )
 
 const (
-	TimeStampLayout = "2006-01-02T15:04:05.000000-07:00"
+	TimeStampLayout          = "2006-01-02T15:04:05.000000-07:00"
+	TimeLayoutYYYYMMDDHHMMSS = "20060102150405"
+	DefaultThresholdLines    = 10
+	DefaultThresholdTime     = 10
+	DefaultThresholdBytes    = 500
 )
+
+type Thresholds struct {
+	Lines int
+	Time  int
+	Bytes int
+}
 
 func extractExitCode(err error) int {
 	if err != nil {
@@ -134,4 +146,94 @@ func RunCmd(command string) (output string, exitCode int, startedAt time.Time, f
 	log.Debugf("")
 	log.Infof("Exit Code: %d", exitCode)
 	return
+}
+
+func RunContinuousReportRun(fn func(chunk string) error, cmdArg string, thresholds Thresholds) error {
+	log.Debug("RunContinuousReportRun")
+
+	// command thresholds flags
+	if !(thresholds.Lines > 0) {
+		thresholds.Lines = DefaultThresholdLines
+	}
+	if !(thresholds.Time > 0) {
+		thresholds.Time = DefaultThresholdTime
+	}
+	if !(thresholds.Bytes > 0) {
+		thresholds.Bytes = DefaultThresholdBytes
+	}
+	log.Debug("Threshold", thresholds)
+
+	// Saves script/command in a temp file
+	cmdFileName := strings.Join([]string{time.Now().Format(TimeLayoutYYYYMMDDHHMMSS), "_", RandomString(10)}, "")
+	if runtime.GOOS == "windows" {
+		cmdFileName = strings.Join([]string{cmdFileName, ".bat"}, "")
+	}
+
+	// Writes content
+	if err := ioutil.WriteFile(cmdFileName, []byte(cmdArg), 0600); err != nil {
+		log.Fatalf("Error creating temp file : ", err)
+	}
+
+	// Creates command
+	var newCmd *exec.Cmd
+	if runtime.GOOS == "windows" {
+		newCmd = exec.Command("cmd", "/C", cmdFileName)
+	} else {
+		newCmd = exec.Command("/bin/sh", cmdFileName)
+	}
+
+	// Removes temp file
+	defer func() {
+		err := os.Remove(cmdFileName)
+		if err != nil {
+			log.Warn("Temp file cannot be removed", err.Error())
+		}
+	}()
+
+	// Gets the pipe command
+	stdout, err := newCmd.StdoutPipe()
+	if err != nil {
+		return fmt.Errorf("cannot get pipe command %v", err)
+	}
+	log.Info("==> Executing: ", strings.Join(newCmd.Args, " "))
+
+	// Start command asynchronously
+	if err = newCmd.Start(); err != nil {
+		return fmt.Errorf("cannot start the specified command %v", err)
+	}
+
+	chunk := ""
+	nLines, nTime, nBytes := 0, 0, 0
+	timeStart := time.Now()
+
+	scanner := bufio.NewScanner(bufio.NewReader(stdout))
+	for scanner.Scan() {
+		chunk = strings.Join([]string{chunk, scanner.Text(), "\n"}, "")
+		nLines++
+		nTime = int(time.Now().Sub(timeStart).Seconds())
+		nBytes = len(chunk)
+		if nLines >= thresholds.Lines || nTime >= thresholds.Time || nBytes >= thresholds.Bytes {
+			if err := fn(chunk); err != nil {
+				nLines, nTime = 0, 0
+				timeStart = time.Now()
+			} else {
+				chunk = ""
+				nLines, nTime, nBytes = 0, 0, 0
+				timeStart = time.Now()
+			}
+		}
+	}
+
+	if err := scanner.Err(); err != nil {
+		log.Error("==> Error: ", err.Error())
+		chunk = strings.Join([]string{chunk, err.Error()}, "")
+	}
+
+	if len(chunk) > 0 {
+		log.Debug("Sending the last pending chunk")
+		if err := fn(chunk); err != nil {
+			return err
+		}
+	}
+	return nil
 }
